Improve infinite scroll content swapping
This also revisits the semantics of the topic page and its CSS classes.

The JS filter module was simplified and uses the default htmx behavior
to swap the innerHTML of the bookshelf navigation.

Finally, I got rid of the "recently-added" partial in favor of the
existing "cover-list" partial. This will serve the goal of #25, by
decreasing the book cover maintenance surface.

The font-size of the cover captions was decreased and a small bug,
making it not using the whole width was fixed.

Closes #31